Use minicore
